Updated imports in safe pilco examples. renamed safe-pilco-extension to safe_pilco_extension
